整理

+++ source/bsys/file/file_pack_makethread.cpp
@@ -147,7 +147,7 @@ namespace NBsys{namespace NFile
 					//各ファイル名文字数。
 					s64 t_filename_length_offset = t_offset;
 					sharedptr<u16> t_filename_length(new u16[t_all_count],default_delete<u16[]>());
-					for(s32 ii=0;ii<static_cast<s32>(t_filename_pack_short_list.size());ii++){
+					for(u32 ii=0;ii<static_cast<u32>(t_filename_pack_short_list.size());ii++){
 						t_filename_length.get()[ii] = static_cast<u16>(t_filename_pack_short_list[ii].length());
 					}
 					t_offset += sizeof(u16) * t_all_count;
@@ -155,8 +155,8 @@ namespace NBsys{namespace NFile
 					//各ファイル名をまとめたもの。
 					s64 t_all_filename_offset = t_offset;
 					STLVector<wchar>::Type t_all_filename;
-					for(s32 ii=0;ii<static_cast<s32>(t_filename_pack_short_list.size());ii++){
-						for(s32 jj=0;jj<static_cast<s32>(t_filename_pack_short_list[ii].length());jj++){
+					for(u32 ii=0;ii<static_cast<u32>(t_filename_pack_short_list.size());ii++){
+						for(u32 jj=0;jj<static_cast<u32>(t_filename_pack_short_list[ii].length());jj++){
 							t_all_filename.push_back(t_filename_pack_short_list[ii][jj]);
 						}
 						t_all_filename.push_back(0x0000);
@@ -178,18 +178,18 @@ namespace NBsys{namespace NFile
 						}
 
 						//各ファイル名文字サイズ。
-						if(t_filehandle_write.Write(reinterpret_cast<u8*>(t_filename_length.get()),sizeof(u16) * t_all_count,t_filename_length_offset) == false){
+						if(t_filehandle_write.Write(reinterpret_cast<u8*>(t_filename_length.get()),static_cast<s32>(sizeof(u16) * t_all_count),t_filename_length_offset) == false){
 							ASSERT(0);
 						}
 
 						//各ファイル名をひとまとめにしたもの。
-						if(t_filehandle_write.Write(reinterpret_cast<u8*>(&t_all_filename[0]),sizeof(wchar) * t_all_filename.size(),t_all_filename_offset) == false){
+						if(t_filehandle_write.Write(reinterpret_cast<u8*>(&t_all_filename[0]),static_cast<s32>(sizeof(wchar) * t_all_filename.size()),t_all_filename_offset) == false){
 							ASSERT(0);
 						}
 					}
 
 					{
-						for(s32 ii=0;ii<static_cast<s32>(t_filename_fullpath_list.size());ii++){
+						for(u32 ii=0;ii<static_cast<s32>(t_filename_fullpath_list.size());ii++){
 							FileHandle t_filehandle_read;
 							t_filehandle_read.ReadOpen(Path::Name(t_filename_fullpath_list[ii]));
 
@@ -198,7 +198,7 @@ namespace NBsys{namespace NFile
 							//ファイルサイズ。
 							t_file_size.get()[ii] = static_cast<u32>(t_filesize_read);
 
-							sharedptr<u8> t_filedata(new u8[static_cast<s32>(t_filesize_read)],default_delete<u8[]>());
+							sharedptr<u8> t_filedata(new u8[static_cast<u32>(t_filesize_read)],default_delete<u8[]>());
 
 							//読み込み。
 							if(t_filehandle_read.Read(t_filedata.get(),t_filesize_read,0) == false){
@@ -218,7 +218,7 @@ namespace NBsys{namespace NFile
 
 					{
 						//各ファイルサイズ。
-						if(t_filehandle_write.Write(reinterpret_cast<u8*>(t_file_size.get()),sizeof(u32) * t_all_count,t_file_size_offset) == false){
+						if(t_filehandle_write.Write(reinterpret_cast<u8*>(t_file_size.get()),static_cast<s32>(sizeof(u32) * t_all_count),t_file_size_offset) == false){
 							ASSERT(0);
 						}
 					}
